% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{deleteAll,neo4jService-method}
\alias{deleteAll,neo4jService-method}
\title{permanently deletes all nodes (and thus all edges)}
\usage{
\S4method{deleteAll}{neo4jService}(obj, confirm = TRUE)
}
\arguments{
\item{obj}{new4jService object}

\item{confirm}{logical, ask user to confirm}
}
\value{
nothing
}
\description{
detaches and deletes all nodes, and thus all edges
}
