% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{fullGraph,neo4jService-method}
\alias{fullGraph,neo4jService-method}
\title{get a 'raw' query result, all edges}
\usage{
\S4method{fullGraph}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
one or more data.frames
}
\description{
returns all the relationships (edges) between all nodes
}
