% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getEdgeCount,neo4jService-method}
\alias{getEdgeCount,neo4jService-method}
\title{how many edges in the current graph?}
\usage{
\S4method{getEdgeCount}{neo4jService}(obj, directed = TRUE)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
the edge count
}
\description{
edge count returned, twice as many when undirected
}
