% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getEdgeTable,neo4jService-method}
\alias{getEdgeTable,neo4jService-method}
\title{return a reusable data.frame listing edges and their properties}
\usage{
\S4method{getEdgeTable}{neo4jService}(obj, directed = TRUE)
}
\arguments{
\item{obj}{new4jService object}

\item{directed}{logical, default TRUE}
}
\value{
a data.frame
}
\description{
edges and their properties
}
