% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getEdgeTypeDistribution,neo4jService-method}
\alias{getEdgeTypeDistribution,neo4jService-method}
\title{getEdgeTypeDistribution}
\usage{
\S4method{getEdgeTypeDistribution}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
a data.frame
}
\description{
report how many nodes are annotated to each label
}
