% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getEdgeTypes,neo4jService-method}
\alias{getEdgeTypes,neo4jService-method}
\title{getEdgeTypes}
\usage{
\S4method{getEdgeTypes}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
character vector
}
\description{
rerutn a sorted uniqued list of edge types
}
