% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getNodeAndEdgeTables,neo4jService-method}
\alias{getNodeAndEdgeTables,neo4jService-method}
\title{return a list of two data.frames, nodes and edges, suitable for rendering in RCyjs}
\usage{
\S4method{getNodeAndEdgeTables}{neo4jService}(obj, query = NA)
}
\arguments{
\item{obj}{neo4jService object}

\item{query}{character string, a valid neo4j query}
}
\value{
a list of two data.frames
}
\description{
transform the neo4j data structures into a two base R data.frames
}
