% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getNodeCount,neo4jService-method}
\alias{getNodeCount,neo4jService-method}
\title{how many nodes in the current graph?}
\usage{
\S4method{getNodeCount}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
the node count
}
\description{
node count returned
}
