% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getNodeLabels,neo4jService-method}
\alias{getNodeLabels,neo4jService-method}
\title{labels seem to be types, categories}
\usage{
\S4method{getNodeLabels}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
character vector
}
\description{
rerutn a sorted uniqued list of label (category) names
}
