% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{getNodeTable,neo4jService-method}
\alias{getNodeTable,neo4jService-method}
\title{return a reusable data.frame listing nodes and their properties}
\usage{
\S4method{getNodeTable}{neo4jService}(obj)
}
\arguments{
\item{obj}{new4jService object}
}
\value{
a data.frame
}
\description{
nodes and their properties
}
