% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\docType{class}
\name{neo4jService-class}
\alias{neo4jService-class}
\alias{.neo4jService}
\alias{neo4jService}
\title{neo4jService-class}
\usage{
neo4jService(host, port, user, password, quiet = TRUE)
}
\arguments{
\item{host}{character string}

\item{port}{numeric}

\item{user}{character string}

\item{password}{character string}

\item{quiet}{logical, default TRUE}
}
\value{
An object of the neo4jService class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
