% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{query.raw,neo4jService-method}
\alias{query.raw,neo4jService-method}
\title{send a query to the database, return it without interpretation or reformatting}
\usage{
\S4method{query.raw}{neo4jService}(obj, s)
}
\arguments{
\item{obj}{new4jService object}

\item{s}{character string}
}
\value{
one or more data.frames
}
\description{
send a well-formed "cypher" query
}
