% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{runCypherFile,neo4jService-method}
\alias{runCypherFile,neo4jService-method}
\title{execute all the cypher commands in the named file}
\usage{
\S4method{runCypherFile}{neo4jService}(obj, filename)
}
\arguments{
\item{obj}{filename character}
}
\value{
not sure...
}
\description{
execute all the cypher commands in the named file
}
