% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4jService.R
\name{shortestPath,neo4jService-method}
\alias{shortestPath,neo4jService-method}
\title{dijkstra's shortest path}
\usage{
\S4method{shortestPath}{neo4jService}(obj, sourceNode, targetNode)
}
\arguments{
\item{obj}{new4jService object}

\item{sourceNode}{character string}

\item{targetNode}{character string}
}
\value{
a list
}
\description{
return the shortest path between two nodes
}
